+++ convey/args_controller.py
@@ -7,7 +7,7 @@ from typing import Annotated, Any, List, Optional, Tuple
 
 from mininterface import run
 from mininterface.interfaces import TextInterface
-from mininterface.types import _BlankTrue, _BlankTrueString, _BlankStr
+from mininterface.tag.flag import BlankTrue, Blank
 from tyro.conf import (FlagConversionOff, OmitArgPrefixes, Positional,
                        UseAppendAction, arg)
 from tyro.extras import get_parser as get_tyro_parser
@@ -36,28 +36,26 @@ class IO:
     file_or_input: Positional[Optional[str | Path]] = None
     """File name to be parsed or input text. In nothing is given, user will input data through stdin."""
 
-    file: _BlankStr = None
-    """Treat <file_or_input> parameter as a file, never as an input"""
-    # group.add_argument('--file', help="Treat <file_or_input> parameter as a file, never as an input",
-    #                    action="store_true")
+    file: Optional[str] = None
+    """Parse file. (Instead of <file_or_input> parameter.)"""
 
-    input: Annotated[_BlankStr, arg(aliases=["-i"])] = ""
-    """Treat <file_or_input> parameter as an input text, not a file name"""
+    input: Annotated[Optional[str], arg(aliases=["-i"])] = ""
+    """Parse input text. (Instead of <file_or_input> parameter.)"""
 
-    output: Annotated[_BlankTrueString, arg(metavar="blank/FILENAME")] = None
+    output: Annotated[Blank[str], arg(metavar="blank|FILENAME")] = None
     """Save output to this file.
     If left blank, pass output to STDOUT.
     If omitted, a filename will be produced automatically.
     May be combined with --headless.
     """
 
-    single_query: Annotated[_BlankTrue, arg(aliases=["-S"])] = None
+    single_query: Annotated[BlankTrue, arg(aliases=["-S"])] = None
     """Consider the input as a single value, not a CSV."""
 
-    single_detect: _BlankTrue = None
+    single_detect: BlankTrue = None
     """Consider the input as a single value, not a CSV, and just print out possible types of the input."""
 
-    csv_processing: Annotated[_BlankTrue, arg(aliases=["-C"])] = None
+    csv_processing: Annotated[BlankTrue, arg(aliases=["-C"])] = None
     """Consider the input as a CSV, not a single."""
 
 
@@ -65,28 +63,28 @@ class IO:
 class CLI:
     """CLI experience"""
 
-    debug: _BlankTrue = None
+    debug: BlankTrue = None
     """Development only: increases verbosity and gets the prompt in the case of an exception."""
 
-    crash_post_mortem: _BlankTrue = None
+    crash_post_mortem: BlankTrue = None
     """Get prompt if program crashes"""
 
-    verbose: Annotated[_BlankTrue, arg(aliases=["-v"])] = None
+    verbose: Annotated[BlankTrue, arg(aliases=["-v"])] = None
     """Sets the verbosity to see DEBUG messages."""
 
-    quiet: Annotated[_BlankTrue, arg(aliases=["-q"])] = None
+    quiet: Annotated[BlankTrue, arg(aliases=["-q"])] = None
     """Sets the verbosity to see WARNINGs and ERRORs only.
     Prints out the least information possible.
     (Ex: if checking single value outputs a single word, prints out just that.)"""
 
-    yes: Annotated[_BlankTrue, arg(aliases=["-y"])] = None
+    yes: Annotated[BlankTrue, arg(aliases=["-y"])] = None
     """Assume non-interactive mode and the default answer to questions.
     Will not send e-mails unless --send is on too."""
 
-    headless: Annotated[_BlankTrue, arg(aliases=["-H"])] = None
+    headless: Annotated[BlankTrue, arg(aliases=["-H"])] = None
     """Launch program in a headless mode which imposes --yes and --quiet. No menu is shown."""
 
-    compute_preview: _BlankTrue = None
+    compute_preview: BlankTrue = None
     """When adding new columns, show few first computed values."""
 
 
@@ -98,42 +96,42 @@ class Environment:
     File: config (default)/uwsgi/template/template_abroad
     Mode: 1 terminal / 2 GUI / 3 try both (default)"""
 
-    show_uml: _BlankTrueString = None
+    show_uml: Blank[int] = None
     """Show UML of fields and methods and exit.
     Methods that are currently disabled via flags or config file are grayed out.
      * FLAGs:
         * +1 to gray out disabled fields/methods
         * +2 to include usual field names"""
 
-    get_autocompletion: _BlankTrue = None
+    get_autocompletion: BlankTrue = None
     """Get bash autocompletion."""
 
-    version: Annotated[_BlankTrue, arg(help=f"""Show the version number (which is currently {
+    version: Annotated[BlankTrue, arg(help=f"""Show the version number (which is currently {
         __version__}).""")] = None
 
 
 @dataclass
 class Processing:
     """ Processing """
-    threads: Annotated[_BlankTrueString, arg(metavar="blank/false/auto/INT")] = None
+    threads: Annotated[Blank[str], arg(metavar="blank/False/auto/INT")] = None
     """Set the thread processing number."""
 
-    fresh: Annotated[_BlankTrue, arg(aliases=["-F"])] = None
+    fresh: Annotated[BlankTrue, arg(aliases=["-F"])] = None
     """Do not attempt to load any previous settings / results.
     Do not load convey's global WHOIS cache.
     (But merge WHOIS results in there afterwards.)"""
 
-    reprocess: Annotated[_BlankTrue, arg(aliases=["-R"])] = None
+    reprocess: Annotated[BlankTrue, arg(aliases=["-R"])] = None
     """Do not attempt to load any previous settings / results.
     But load convey's global WHOIS cache."""
 
-    server: _BlankTrue = None
+    server: BlankTrue = None
     """Launches simple web server."""
 
-    daemon: Annotated[_BlankTrueString, arg(metavar="start/restart/stop/status/server")] = None
+    daemon: Annotated[Blank[str], arg(metavar="start/restart/stop/status/server")] = None
     """Run a UNIX socket daemon to speed up single query requests.
-      * 1/true/on – allow using the daemon
-      * 0/false/off – do not use the daemon
+      * True – allow using the daemon
+      * False – do not use the daemon
       * start – start the daemon and exit
       * stop – stop the daemon and exit
       * status – print out the status of the daemon
@@ -151,7 +149,7 @@ class CSVDialect:
     quote_char: str = ""
     """Treat file as having this quoting character."""
 
-    header: _BlankTrue = None
+    header: BlankTrue = None
     """Treat file as having header."""
 
     delimiter_output: str = ""
@@ -160,7 +158,7 @@ class CSVDialect:
     quote_char_output: str = ""
     """Output quoting char."""
 
-    header_output: _BlankTrue = None
+    header_output: BlankTrue = None
     """If false, header is omitted when processing."""
 
 
@@ -217,16 +215,16 @@ class Actions:
 @dataclass
 class EnablingModules:
     """ Enabling modules """
-    whois: _BlankTrue = None
+    whois: BlankTrue = None
     """Allowing Whois module: Leave blank for True or put true/on/1 or false/off/0."""
 
-    nmap: _BlankTrue = None
+    nmap: BlankTrue = None
     """Allowing NMAP module: Leave blank for True or put true/on/1 or false/off/0."""
 
-    dig: _BlankTrue = None
+    dig: BlankTrue = None
     """Allowing DNS DIG module: Leave blank for True or put true/on/1 or false/off/0."""
 
-    web: _BlankTrue = None
+    web: BlankTrue = None
     """Allowing Web module: Leave blank for True or put true/on/1 or false/off/0.
     When single value input contains a web page, we could fetch it and add status (HTTP code) and text fields.
     Text is just mere text, no tags, style, script, or head."""
@@ -235,19 +233,19 @@ class EnablingModules:
 @dataclass
 class FieldComputingOptions:
     """ Field computing options """
-    disable_external: _BlankTrue = None
+    disable_external: BlankTrue = None
     """Disable external function registered in config.ini to be imported."""
 
-    json: _BlankTrue = None
+    json: BlankTrue = None
     """When checking single value, prefer JSON output rather than text."""
 
     user_agent: str = ""
     """Change user agent to be used when scraping a URL."""
 
-    multiple_hostname_ip: _BlankTrue = None
+    multiple_hostname_ip: BlankTrue = None
     """Hostname can be resolved into multiple IP addresses. Duplicate row for each."""
 
-    multiple_cidr_ip: _BlankTrue = None
+    multiple_cidr_ip: BlankTrue = None
     """CIDR can be resolved into multiple IP addresses. Duplicate row for each."""
 
     web_timeout: Annotated[int, arg(metavar="SECONDS")] = 30
@@ -260,38 +258,38 @@ class WhoisModule:
     whois_ttl: Annotated[int, arg(metavar="SECONDS")] = 86400
     """How many seconds will a WHOIS answer cache will be considered fresh."""
 
-    whois_delete: _BlankTrue = None
+    whois_delete: BlankTrue = None
     """Delete convey's global WHOIS cache."""
 
-    whois_delete_unknown: _BlankTrue = None
+    whois_delete_unknown: BlankTrue = None
     """Delete unknown prefixes from convey's global WHOIS cache."""
 
-    whois_reprocessable_unknown: _BlankTrue = None
+    whois_reprocessable_unknown: BlankTrue = None
     """Make unknown lines reprocessable while single file processing, do not leave unknown cells empty."""
 
-    whois_cache: _BlankTrue = None
+    whois_cache: BlankTrue = None
     """Use whois cache."""
 
 
 @dataclass
 class SendingOptions:
     """ Sending options """
-    send: Annotated[_BlankTrueString, arg(metavar="blank/smtp/otrs")] = ""
+    send: Annotated[Blank[str], arg(metavar="blank/smtp/otrs")] = ""
     """Automatically send e-mails when split."""
 
     send_test: Optional[Annotated[tuple[str, str], arg(metavar=("E-MAIL", "TEMPLATE_FILE"))]] = None
     """Display e-mail message that would be generated for given e-mail."""
 
-    jinja: _BlankTrue = None
+    jinja: BlankTrue = None
     """Process e-mail messages with jinja2 templating system."""
 
-    attach_files: _BlankTrue = None
+    attach_files: BlankTrue = None
     """Split files are added as e-mail attachments."""
 
-    attach_paths_from_path_column: _BlankTrue = None
+    attach_paths_from_path_column: BlankTrue = None
     """Files from a column of the Path type are added as e-mail attachments."""
 
-    testing: _BlankTrue = None
+    testing: BlankTrue = None
     """Do not be afraid, e-mail messages will not be sent. They will get forwarded to the testing e-mail."""
 
     subject: str = ""
@@ -306,12 +304,20 @@ class SendingOptions:
 
 @dataclass
 class OTRS:
-    otrs_id: str = ""
+    """ OTRS specific options. We may send all the e-mails by it. """
+
+    enabled: bool = False
+
+    id: str = ""
     """Ticket id"""
-    otrs_cookie: str = ""
+    cookie: str = ""
     """OTRSAgentInterface cookie"""
-    otrs_token: str = ""
+    token: str = ""
     """OTRS challenge token"""
+
+    host: str = "localhost"
+    baseuri: str = "/otrs/index.pl"
+    signkeyid: str = "PGP::Sign::-"
 
 
 @dataclass